refactor : 서치 페이지 리팩토링 - Explore 스타일 파일 Search 스타일 파일로 병합 - 각 페이지 맥락에 맞게 스타일드 컴포넌트 이름 변경해서 사용 - 실시간 검색결과 패널에서 deferredValue, debouncingInput 상위 컴포넌트로 위임
