--- src/components/Explore/Explore.styles.ts
@@ -1,24 +1,0 @@
-import styled from "@emotion/styled";
-
-export const ExplorePageLayout = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-  padding: 20px;
-  gap: 12px;
-  background-color: white;
-`;
-
-export const SearchBarLayout = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: start;
-  gap: 20px;
-  width: 100%;
-  height: 45px;
-  padding: 0 20px;
-  border-radius: 10px;
-  background: white;
-  z-index: 70;
-`;
